multimodal_rag_pipeline/requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.2.0

# Document processing
pypdf>=3.15.0
python-docx>=0.8.11
beautifulsoup4>=4.12.0
pdfplumber>=0.10.0
unstructured>=0.10.0
unstructured-inference>=0.6.0
layoutparser>=0.3.4
detectron2>=0.6
pdf2image>=1.16.0

# Text processing
transformers>=4.30.0
spacy>=3.6.0
nltk>=3.8.0
sentence-transformers>=2.2.2

# Image processing
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0
clip>=1.0
torchvision>=0.15.0

# Audio processing (optional)
librosa>=0.10.0
whisper>=1.0

# Embedding and vector databases
faiss-cpu>=1.7.4  # or faiss-gpu for GPU support
pinecone-client>=2.2.0
weaviate-client>=3.20.0
qdrant-client>=1.5.0
chromadb>=0.4.0

# LLM integration
langchain>=0.0.300
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0.0
anthropic>=0.5.0
huggingface-hub>=0.16.0
accelerate>=0.21.0
bitsandbytes>=0.41.0  # For model quantization

# Multimodal models
open-clip-torch>=2.20.0
flava-pytorch>=0.1.0

# Evaluation
rouge>=1.0.1
bert-score>=0.3.13
pytrec-eval>=0.5

# Utilities
jupyter>=1.0.0
ipywidgets>=8.0.0
python-dotenv>=1.0.0